heic-converter/heic_tool/pyproject.toml
[project]
name = "heic-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.0.0",
    "pillow-heif>=1.1.1",
    "streamlit>=1.52.1",
]
